Extracts recording related logic from JitsiMeetConference to separate class.
